refactor: view 패키지 추가 및 글자 출력 펑션 이동 및 적용, 게임 재시작 방어로직 추가
