+++ src/main/java/view/SystemMessage.java
@@ -1,4 +1,4 @@
-package utils;
+package view;
 
 public class SystemMessage {
     public void getEndMessage(){
